+++ pkg/repository/udmrepo/kopialib/logging/logging.go
@@ -1,3 +1,5 @@
+package logging
+
 /*
 Copyright the Velero contributors.
 
@@ -13,8 +15,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-
-package logging
 
 import (
 	"context"
